+++ src/features/pages/check-email/check-email.steps.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { check_email_steps } from './check-email.module.scss'
 import { TString } from 'types/generics'
-import Spam from 'images/svg/check-email/spam.svg'
-import Typo from 'images/svg/check-email/typo.svg'
-import WorkEmail from 'images/svg/check-email/work-email.svg'
-import Firewalls from 'images/svg/check-email/firewalls.svg'
+import Spam from 'images/svg/check-email/check_email_spam.svg'
+import Typo from 'images/svg/check-email/check_email_typo.svg'
+import WorkEmail from 'images/svg/check-email/check_email_work_email.svg'
+import Firewalls from 'images/svg/check-email/check_email_firewalls.svg'
 import Flex from 'features/components/atoms/flex-box'
 import Typography from 'features/components/atoms/typography'
 import { Localize } from 'components/localization'
