+++ src/features/pages/ctrader-signup/signup-credentials.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { form_style, paragraph_style } from './ctrader-form.module.scss'
 import CtraderCredentialsForm from './ctrader-credentials-form'
 import { Localize } from 'components/localization'
-import EmailIcon from 'images/svg/check-email/email.svg'
+import EmailIcon from 'images/svg/ctrader/ctrader_email.svg'
 import { isBrowser } from 'common/utility'
 import CtraderWrapper from 'features/components/templates/ctrader/ctrader-wrapper'
 import Flex from 'features/components/atoms/flex-box'
