+++ src/pages/careers/_employee-testimonial-carousel.tsx
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types'
 import { Carousel, CarouselProps, Header, Text } from 'components/elements'
 import device from 'themes/device'
 import { SectionContainer, Flex } from 'components/containers'
-import QuoteMark from 'images/svg/careers/quotemark.svg'
+import QuoteMark from 'images/svg/careers/career_quotemark.svg'
 import AhmadImage from 'images/common/careers/ahmad.jpg'
 import AdityaImage from 'images/common/careers/aditya.jpg'
 import GaryImage from 'images/common/careers/gary.jpg'
