+++ src/pages/careers/_our-hiring-process.tsx
@@ -3,10 +3,10 @@ import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import { zoho_url } from 'common/constants'
 // SVG
-import Career1 from 'images/svg/careers/home/home_1.svg'
-import Career2 from 'images/svg/careers/home/home_2.svg'
-import Career3 from 'images/svg/careers/home/home_3.svg'
-import Career4 from 'images/svg/careers/home/home_4.svg'
+import Career1 from 'images/svg/careers/career_1.svg'
+import Career2 from 'images/svg/careers/career_2.svg'
+import Career3 from 'images/svg/careers/career_3.svg'
+import Career4 from 'images/svg/careers/career_4.svg'
 import DownwardsLeft from 'images/svg/careers/career_downwards_left.svg'
 import DownwardsRight from 'images/svg/careers/career_downwards_right.svg'
 import { Container, SectionContainer, Flex } from 'components/containers'
